# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WeSwing",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/M14/Projecte_Final/grup2/server/public/index.php"
        }
    ],
    "paths": {
...
        "/api/seguits": {
            "get": {
                "tags": [
                    "Seguidors"
                ],
                "summary": "Veure tots els usuaris seguits",
                "description": "",
                "operationId": "App\\Http\\Controllers\\SeguitController::getSeguits",
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "500": {
                        "description": "S'ha produit un error"
                    }
                }
            }
        },
        "/api/seguit/{id}": {
            "get": {
                "tags": [
                    "Seguidors"
                ],
                "summary": "Veure el seguidor desitjada.",
                "description": "",
                "operationId": "App\\Http\\Controllers\\SeguitController::getSeguitsId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Seguidor per id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seguidor amb id seleccionat"
                    },
                    "500": {
                        "description": "S'ha produit un error"
                    }
                }
            }
        },
        "/api/seguir": {
            "post": {
                "tags": [
                    "Seguidors"
                ],
                "summary": "Inserir un nou seguidor.",
                "operationId": "App\\Http\\Controllers\\SeguitController::seguir",
                "parameters": [
                    {
                        "name": "seguit_id",
                        "in": "query",
                        "description": "id del seguit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "seguidor_id",
                        "in": "query",
                        "description": "id de la persona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna l'usuari que hem seguit."
                    },
                    "default": {
                        "description": "S'ha produit un error."
                    }
                }
            }
        },
        "/api/tipusballs": {
            "get": {
                "tags": [
                    "Tipus de Bll"
                ],
                "summary": "Veure tots els tipus de ball.",
                "operationId": "App\\Http\\Controllers\\TipusBallController::getTipusBalls",
                "responses": {
                    "200": {
                        "description": "Retorna tots els tipus de ball."
                    },
                    "default": {
                        "description": "S'ha produit un error."
                    }
                }
            }
        },
        "/api/tipusball/{id}}": {
            "get": {
                "tags": [
                    "Tipus de Bll"
                ],
                "summary": "Veure el tipus de ball desitjat.",
                "description": "",
                "operationId": "App\\Http\\Controllers\\TipusBallController::getTipusBall",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tipus de ball per id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ATipus de ball amb id seleccionat"
                    },
                    "500": {
                        "description": "S'ha produit un error"
                    }
                }
            }
        },
        "/api/tipusball/{id}": {
            "post": {
                "tags": [
                    "Tipus de Bll"
                ],
                "summary": "Actualitza el tipus de ball ja existent",
                "description": "",
                "operationId": "updateTipusBall",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id tipus de ball a modificar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "nom",
                        "in": "query",
                        "description": "Nom",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipus de ball modificat"
                    },
                    "500": {
                        "description": "S'ha produit un error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Agrupacions"
                ],
                "summary": "Eliminar un tipus de ball.",
                "description": "",
                "operationId": "App\\Http\\Controllers\\TipusBallController::eliminarTipusBall",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tipus de ball a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipus de ball eliminat"
                    },
                    "500": {
                        "description": "S'ha produit un error"
                    }
                }
            }
        },
        "/api/tipusball": {
            "post": {
                "tags": [
                    "Tipus de Bll"
                ],
                "summary": "Inserir un nou tipus de ball.",
                "operationId": "App\\Http\\Controllers\\TipusBallController::crearTipusBall",
                "parameters": [
                    {
                        "name": "nom",
                        "in": "query",
                        "description": "Nom del tipus de ball",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna el tipus de ball que hem inserit."
                    },
                    "default": {
                        "description": "S'ha produit un error."
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}